+++ spec/support/entities/computer.rb
@@ -1,14 +1,14 @@
-require_relative 'computer_hardware'
-require_relative 'computer_software'
+require_relative 'computer_component/hardware'
+require_relative 'computer_component/software'
 
 class Computer
   include Datamappify::Entity
 
   attribute :brand, String
 
-  attributes_from ComputerHardware
-  attributes_from ComputerSoftware, :prefix_with => :software
-  attributes_from ComputerSoftware, :prefix_with => :game
+  attributes_from ComputerComponent::Hardware
+  attributes_from ComputerComponent::Software, :prefix_with => :software
+  attributes_from ComputerComponent::Software, :prefix_with => :game
 
   validates :brand, :presence => true
 end