+++ spec/support/entities/computer_component/software.rb
@@ -1,0 +1,10 @@
+module ComputerComponent
+  class Software
+    include Datamappify::Entity
+
+    attribute :os, String
+    attribute :vendor, String
+
+    validates_presence_of :os
+  end
+end